Conflicts:
	modules/timeline-avr/app.o
	modules/timeline-avr/app.out
	modules/timeline-avr/helpers.o
	modules/timeline-avr/rom.hex
	modules/timeline-avr/timeline.map
	modules/timeline-avr/timeline.o

+++ modules/timer/timer.c
@@ -35,7 +35,7 @@ ISR(TIMER4_OVF_vect) {
     time_count++;
 }
 
-int GetCurrentTime() {
+short int GetCurrentTime() {
     return time_count;
 }
